feat: move notifications into components
